# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartQueue AI - Intelligent Virtual Queue Management System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  admin: "admin",
  staff: "staff",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  phone?: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  admin: "admin",
  staff: "staff",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type QueueIndustry = (typeof QueueIndustry)[keyof typeof QueueIndustry];

export const QueueIndustry = {
  hospital: "hospital",
  bank: "bank",
  salon: "salon",
  college: "college",
  government: "government",
  retail: "retail",
  other: "other",
} as const;

export type QueueStatus = (typeof QueueStatus)[keyof typeof QueueStatus];

export const QueueStatus = {
  active: "active",
  paused: "paused",
  closed: "closed",
} as const;

export interface Queue {
  id: number;
  name: string;
  description?: string;
  industry: QueueIndustry;
  location: string;
  avgServiceTimeMinutes: number;
  status: QueueStatus;
  maxCapacity?: number;
  currentCount: number;
  estimatedWaitMinutes: number;
  openAt?: string;
  closeAt?: string;
  createdAt: string;
}

export type TokenStatus = (typeof TokenStatus)[keyof typeof TokenStatus];

export const TokenStatus = {
  waiting: "waiting",
  called: "called",
  serving: "serving",
  completed: "completed",
  skipped: "skipped",
  left: "left",
  no_show: "no_show",
} as const;

export type TokenPriority = (typeof TokenPriority)[keyof typeof TokenPriority];

export const TokenPriority = {
  normal: "normal",
  vip: "vip",
  emergency: "emergency",
} as const;

export interface Token {
  id: number;
  queueId: number;
  userId: number;
  tokenNumber: string;
  position: number;
  status: TokenStatus;
  priority: TokenPriority;
  estimatedWaitMinutes: number;
  noShowProbability?: number;
  arrivalConfirmed: boolean;
  joinedAt: string;
  calledAt?: string;
  completedAt?: string;
}

export type QueueDetail = Queue & {
  tokens?: Token[];
  calledToken?: Token;
};

export type CreateQueueBodyIndustry =
  (typeof CreateQueueBodyIndustry)[keyof typeof CreateQueueBodyIndustry];

export const CreateQueueBodyIndustry = {
  hospital: "hospital",
  bank: "bank",
  salon: "salon",
  college: "college",
  government: "government",
  retail: "retail",
  other: "other",
} as const;

export interface CreateQueueBody {
  name: string;
  description?: string;
  industry: CreateQueueBodyIndustry;
  location: string;
  avgServiceTimeMinutes: number;
  maxCapacity?: number;
  openAt?: string;
  closeAt?: string;
}

export type UpdateQueueBodyStatus =
  (typeof UpdateQueueBodyStatus)[keyof typeof UpdateQueueBodyStatus];

export const UpdateQueueBodyStatus = {
  active: "active",
  paused: "paused",
  closed: "closed",
} as const;

export interface UpdateQueueBody {
  name?: string;
  description?: string;
  avgServiceTimeMinutes?: number;
  maxCapacity?: number;
  openAt?: string;
  closeAt?: string;
  status?: UpdateQueueBodyStatus;
}

export interface PauseQueueBody {
  paused: boolean;
}

export type TokenDetailConfidenceLevel =
  (typeof TokenDetailConfidenceLevel)[keyof typeof TokenDetailConfidenceLevel];

export const TokenDetailConfidenceLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type TokenDetail = Token & {
  queue?: Queue;
  user?: User;
  qrCode?: string;
  confidenceLevel?: TokenDetailConfidenceLevel;
};

export type JoinQueueBodyPriority =
  (typeof JoinQueueBodyPriority)[keyof typeof JoinQueueBodyPriority];

export const JoinQueueBodyPriority = {
  normal: "normal",
  vip: "vip",
  emergency: "emergency",
} as const;

export interface JoinQueueBody {
  priority?: JoinQueueBodyPriority;
  notes?: string;
}

export interface LeaveQueueBody {
  reason?: string;
}

export interface AnalyticsSummary {
  totalTokensServed: number;
  avgWaitTimeMinutes: number;
  noShowRate: number;
  peakHour?: string;
  activeQueues: number;
  totalQueues: number;
  tokensToday: number;
  satisfactionScore?: number;
}

export interface DailyStat {
  date: string;
  count: number;
  avgWait: number;
  noShows: number;
}

export interface QueueAnalytics {
  queueId: number;
  queueName: string;
  totalServed: number;
  avgWaitTime: number;
  noShowRate: number;
  dailyStats: DailyStat[];
  topPeakHour?: string;
}

export interface PeakHourData {
  dayOfWeek: number;
  hour: number;
  avgCount: number;
}

export interface DisplayData {
  queueName: string;
  currentlyServing: string;
  waitingCount: number;
  recentlyCalled: string[];
  estimatedWaitMinutes: number;
  status: string;
}

export type ListQueuesParams = {
  industry?: string;
  active?: boolean;
};

export type ListTokensParams = {
  status?: string;
};

export type GetAnalyticsSummaryParams = {
  days?: number;
};

export type GetQueueAnalyticsParams = {
  days?: number;
};

export type GetPeakHoursParams = {
  queueId?: number;
  days?: number;
};